[REF] Add topography, implants.cortex, and models.cortex submodules (#518)
* [MNT] Update requirements.txt (#507)

* [DOC] Fix gallery thumbnail images (#510)

* add topography module

* fix imports

* refactor tests for topography module

* doc

* add epmty submodules for implants.cortex and models.cortex
