+++ pulse2percept/implants/cortex/__init__.py
@@ -1,0 +1,13 @@
+"""
+Different cortical prosthetic implants such as TODO.
+
+.. autosummary::
+    : toctree: _api
+
+.. seealso::
+
+    *  :ref:`Basic Concepts > Visual Prostheses <topics-implants>`
+"""
+
+__all__ = [
+]
