+++ pulse2percept/models/cortex/__init__.py
@@ -1,0 +1,17 @@
+"""Computational models of the cortex, such as phosphene and neural response models.
+
+.. autosummary::
+    :toctree: _api
+
+    base
+
+.. seealso::
+
+    *  :ref:`Basic Concepts > Computational Models <topics-models>`
+
+"""
+from .base import * # TODO
+
+
+__all__ = [
+]
