+++ pulse2percept/models/tests/test_thompson2003.py
@@ -11,7 +11,7 @@ import matplotlib.pyplot as plt
 from pulse2percept.implants import ArgusI, ArgusII
 from pulse2percept.percepts import Percept
 from pulse2percept.models import (Thompson2003Spatial, Thompson2003Model)
-from pulse2percept.utils import Curcio1990Map, Watson2014DisplaceMap
+from pulse2percept.topography import Curcio1990Map, Watson2014DisplaceMap
 from pulse2percept.utils.testing import assert_warns_msg
