+++ pulse2percept/viz/argus.py
@@ -12,8 +12,9 @@ from pkg_resources import resource_filename
 
 from ..implants import ArgusI, ArgusII
 from ..models import AxonMapModel
-from ..utils import Watson2014Map, scale_image, center_image
+from ..utils import scale_image, center_image
 from ..utils.constants import ZORDER
+from ..topography import Watson2014Map
 
 PATH_ARGUS1 = resource_filename('pulse2percept', 'viz/data/argus1.png')
 PATH_ARGUS2 = resource_filename('pulse2percept', 'viz/data/argus2.png')
